refactor(themes): change theme logic to a separate hook according to SRP
