+++ apps/web/package.json
@@ -9,29 +9,6 @@
     "lint": "next lint",
     "lint:fix": "next lint --fix"
   },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "pages/**/*.{js,jsx,ts,tsx}": [
-      "npm run lint:fix",
-      "git add"
-    ],
-    "components/**/*.{js,jsx,ts,tsx}": [
-      "npm run lint:fix",
-      "git add"
-    ],
-    "styles/**/*.{css,scss,sass}": [
-      "npm run lint:fix",
-      "git add"
-    ],
-    "**/*.{js,jsx,ts,tsx,json,css,scss,sass,md}": [
-      "prettier --write",
-      "git add"
-    ]
-  },
   "dependencies": {
     "@codelife-ui/react": "*",
     "@codelife-ui/tokens": "*",
@@ -60,7 +37,6 @@
     "autoprefixer": "^10.4.14",
     "eslint": "^8.38.0",
     "eslint-plugin-tailwindcss": "^3.11.0",
-    "husky": "^8.0.3",
     "postcss": "^8.4.21",
     "prettier": "^2.8.7",
     "tailwindcss": "^3.3.1"
