+++ apps/web/src/pages/Layout.tsx
@@ -2,45 +2,26 @@
 import Head from 'next/head'
 import Footer from '@/components/Footer'
 import { Box, Switch, Text } from '@codelife-ui/react'
-import { useTheme } from 'next-themes'
 import React, { useEffect, useState } from 'react'
 import { FaMoon, FaSun } from 'react-icons/fa'
 import { useRouter } from 'next/router'
 import { useTranslation } from 'react-i18next'
+import { useTheme } from '@/hooks/useTheme'
 interface LayoutProps {}
 const Layout: React.FC<React.PropsWithChildren<LayoutProps>> = ({
   children,
 }) => {
   const { locale } = useRouter()
   const { i18n } = useTranslation()
-  const { theme, systemTheme, setTheme } = useTheme()
   const [hydrated, setHydrated] = useState(false)
+  const { theme, toggleTheme } = useTheme()
   useEffect(() => {
     setHydrated(true)
   }, [])
   useEffect(() => {
     i18n.changeLanguage(locale)
   }, [locale, i18n])
-  useEffect(() => {
-    const storagedTheme = localStorage.getItem('codelife-theme')
-    if (!systemTheme || theme === systemTheme || storagedTheme !== systemTheme)
-      return
-    setTheme(systemTheme)
-  }, [systemTheme, setTheme, theme])
-  useEffect(() => {
-    const storagedTheme = localStorage.getItem('codelife-theme')
-    if (storagedTheme && storagedTheme !== theme) {
-      setTheme(storagedTheme)
-    }
-  }, [])
-  useEffect(() => {
-    if (!theme) return
-    localStorage.setItem('codelife-theme', theme)
-  }, [theme])
-  const toggleTheme = () => {
-    const currentTheme = theme === 'dark' ? 'light' : 'dark'
-    setTheme(currentTheme)
-  }
+
   return (
     <>
       <Head>
